Correzione Utente Num cell

+++ src/amazon/DBConnection.java
@@ -385,7 +385,7 @@ public class DBConnection {
     * @param cellulare
     * @throws SQLException 
     */
-   public static void creaUtente(String nome, String cognome, String mail, String password, int cellulare) throws SQLException
+   public static void creaUtente(String nome, String cognome, String mail, String password, long cellulare) throws SQLException
    {
        PreparedStatement pstmt; //Statement inserimento nuova riga in ordini
        
@@ -394,7 +394,7 @@ public class DBConnection {
        pstmt.setString(2, cognome);
        pstmt.setString(3, mail);
        pstmt.setString(4, password);
-       pstmt.setInt(5, cellulare);
+       pstmt.setLong(5, cellulare);
        
        pstmt.executeUpdate();
    }
@@ -409,7 +409,7 @@ public class DBConnection {
     * @param cellulare
     * @throws SQLException 
     */
-   public static void aggiornaUtente(int id, String nome, String cognome, String mail, String password, int cellulare) throws SQLException {
+   public static void aggiornaUtente(int id, String nome, String cognome, String mail, String password, long cellulare) throws SQLException {
        PreparedStatement pstmt; //Statement inserimento nuova riga in ordini
        
        pstmt = conn.prepareStatement("UPDATE UTENTI SET NOME = ?, COGNOME = ?, EMAIL = ?, PSW = ?, NUMCELLULARE = ? WHERE UTENTE_ID = ?");
@@ -417,7 +417,7 @@ public class DBConnection {
        pstmt.setString(2, cognome);
        pstmt.setString(3, mail);
        pstmt.setString(4, password);
-       pstmt.setInt(5, cellulare);
+       pstmt.setLong(5, cellulare);
        pstmt.setInt(6, id);
        
        pstmt.executeUpdate();
